products are now being added to cart with discount and quantity

+++ POS/Model/Discount+CoreDataProperties.swift
@@ -2,19 +2,22 @@
 //  Discount+CoreDataProperties.swift
 //  POS
 //
-//  Created by Amit  Singh on 9/18/18.
+//  Created by Amit  Singh on 9/19/18.
 //  Copyright © 2018 singhamit089. All rights reserved.
 //
 //
 
-import CoreData
 import Foundation
+import CoreData
+
 
 extension Discount {
+
     @nonobjc public class func fetchRequest() -> NSFetchRequest<Discount> {
         return NSFetchRequest<Discount>(entityName: "Discount")
     }
 
     @NSManaged public var name: String?
-    @NSManaged public var value: Int32
+    @NSManaged public var value: Double
+
 }
